requirements.txt
autogen==1.0.16
ipython==8.12.3
jsonschema==4.22.0
matplotlib==3.8.4
networkx==3.3
PyYAML==6.0.1
PyYAML==6.0.1
attrs==23.2.0
pytest==8.2.2
pytest-asyncio==0.23.7
